ts-730 revert renaming handlers
