+++ tstreams-transaction-server/src/main/scala/com/bwsw/tstreamstransactionserver/netty/server/singleNode/hanlder/commitLog/GetCommitLogOffsetsHandler.scala
@@ -16,25 +16,28 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.data
+package com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.commitLog
 
-import com.bwsw.tstreamstransactionserver.netty.server.TransactionServer
+import com.bwsw.tstreamstransactionserver.netty.server.commitLogService.ScheduledCommitLog
 import com.bwsw.tstreamstransactionserver.netty.server.handler.PredefinedContextHandler
-import com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.data.TransactionDataPutHandler._
+import com.bwsw.tstreamstransactionserver.netty.server.singleNode.commitLogService.CommitLogService
 import com.bwsw.tstreamstransactionserver.netty.{Protocol, RequestMessage}
-import com.bwsw.tstreamstransactionserver.rpc.{ServerException, TransactionService}
-import com.bwsw.tstreamstransactionserver.tracing.ServerTracer.tracer
+import com.bwsw.tstreamstransactionserver.rpc.{CommitLogInfo, ServerException, TransactionService}
+
+import com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.commitLog.GetCommitLogOffsetsHandler._
+
+
 import io.netty.channel.ChannelHandlerContext
 
 import scala.concurrent.ExecutionContext
 
-
-private object TransactionDataPutHandler {
-  val descriptor = Protocol.PutTransactionData
+private object GetCommitLogOffsetsHandler {
+  val descriptor = Protocol.GetCommitLogOffsets
 }
 
-class TransactionDataPutHandler(server: TransactionServer,
-                                context: ExecutionContext)
+class GetCommitLogOffsetsHandler(commitLogService: CommitLogService,
+                                 scheduledCommitLog: ScheduledCommitLog,
+                                 context: ExecutionContext)
   extends PredefinedContextHandler(
     descriptor.methodID,
     descriptor.name,
@@ -42,37 +45,29 @@ class TransactionDataPutHandler(server: TransactionServer,
 
   override def createErrorResponse(message: String): Array[Byte] = {
     descriptor.encodeResponse(
-      TransactionService.PutTransactionData.Result(
+      TransactionService.GetCommitLogOffsets.Result(
         None,
-        Some(ServerException(message)
-        )
+        Some(ServerException(message))
       )
     )
   }
 
-  override protected def fireAndForget(message: RequestMessage): Unit =
-    tracer.withTracing(message, name = getClass.getName + ".fireAndForget")(process(message))
-
-  private def process(message: RequestMessage) = {
-    tracer.withTracing(message, name = getClass.getName + ".process") {
-      val args = descriptor.decodeRequest(message.body)
-
-      server.putTransactionData(
-        args.streamID,
-        args.partition,
-        args.transaction,
-        args.data,
-        args.from
-      )
-    }
-  }
+  override protected def fireAndForget(message: RequestMessage): Unit = {}
 
   override protected def getResponse(message: RequestMessage,
                                      ctx: ChannelHandlerContext): Array[Byte] = {
-    tracer.withTracing(message, name = getClass.getName + ".getResponse") {
-      descriptor.encodeResponse(
-        TransactionService.PutTransactionData.Result(
-          Some(process(message))))
-    }
+    val response = descriptor.encodeResponse(
+      TransactionService.GetCommitLogOffsets.Result(
+        Some(process(message.body))
+      )
+    )
+    response
+  }
+
+  private def process(requestBody: Array[Byte]) = {
+    CommitLogInfo(
+      commitLogService.getLastProcessedCommitLogFileID,
+      scheduledCommitLog.currentCommitLogFile
+    )
   }
 }
