+++ tstreams-transaction-server/src/main/scala/com/bwsw/tstreamstransactionserver/netty/server/singleNode/hanlder/data/PutSimpleTransactionAndDataHandler.scala
@@ -21,7 +21,7 @@ package com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.data
 import com.bwsw.tstreamstransactionserver.netty.server.batch.Frame
 import com.bwsw.tstreamstransactionserver.netty.server.commitLogService.ScheduledCommitLog
 import com.bwsw.tstreamstransactionserver.netty.server.handler.ArgsDependentContextHandler
-import com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.data.SimpleTransactionAndDataPutHandler._
+import com.bwsw.tstreamstransactionserver.netty.server.singleNode.hanlder.data.PutSimpleTransactionAndDataHandler._
 import com.bwsw.tstreamstransactionserver.netty.server.subscriber.OpenedTransactionNotifier
 import com.bwsw.tstreamstransactionserver.netty.server.{OrderedExecutionContextPool, TransactionServer}
 import com.bwsw.tstreamstransactionserver.netty.{Protocol, RequestMessage}
@@ -32,11 +32,11 @@ import io.netty.channel.ChannelHandlerContext
 
 import scala.concurrent.{ExecutionContext, Future}
 
-private object SimpleTransactionAndDataPutHandler {
+private object PutSimpleTransactionAndDataHandler {
   val descriptor = Protocol.PutSimpleTransactionAndData
 }
 
-class SimpleTransactionAndDataPutHandler(server: TransactionServer,
+class PutSimpleTransactionAndDataHandler(server: TransactionServer,
                                          scheduledCommitLog: ScheduledCommitLog,
                                          notifier: OpenedTransactionNotifier,
                                          authOptions: AuthenticationOptions,
@@ -150,7 +150,6 @@ class SimpleTransactionAndDataPutHandler(server: TransactionServer,
         isNotReliable = false
       )
     }(context)
-
     (result, context)
   }
 }
